fix book detail

+++ app/src/main/java/com/example/smartlamp/model/BookShowModel.kt
@@ -4,5 +4,5 @@ data class BookShowModel(
     var image: String,
     val name: String,
     val rate: Double,
-    val bookId: Int
+    val bookId: Long
 )
